+++ packages/core/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.14.0](https://github.com/pattern-lab/patternlab-node/tree/master/packages/core/compare/v5.13.3...v5.14.0) (2021-01-12)
+
+**Note:** Version bump only for package @pattern-lab/core
+
+
+
+
+
+## [5.13.3](https://github.com/pattern-lab/patternlab-node/tree/master/packages/core/compare/v5.13.2...v5.13.3) (2020-12-17)
+
+**Note:** Version bump only for package @pattern-lab/core
+
+
+
+
+
+## [5.13.2](https://github.com/pattern-lab/patternlab-node/tree/master/packages/core/compare/v5.13.1...v5.13.2) (2020-11-12)
+
+**Note:** Version bump only for package @pattern-lab/core
+
+
+
+
+
+## [5.13.1](https://github.com/pattern-lab/patternlab-node/tree/master/packages/core/compare/v5.13.0...v5.13.1) (2020-09-06)
+
+**Note:** Version bump only for package @pattern-lab/core
+
+
+
+
+
 # [5.13.0](https://github.com/pattern-lab/patternlab-node/tree/master/packages/core/compare/v5.12.0...v5.13.0) (2020-08-26)
 
 **Note:** Version bump only for package @pattern-lab/core